fix: test, and add env testing
